+++ tests/testthat/helper-modules.R
@@ -3,7 +3,7 @@ import_stdlib_modules <- function(engine, modules, env = NULL) {
     env <- engine$get_env()
   }
   for (module in modules) {
-    exprs <- engine$read(sprintf("(import %s)", module))
+    exprs <- engine$read(sprintf("(import %s :refer :all)", module))
     engine$eval(exprs[[1]], env = env)
   }
   env